add: auth package, fix: delete json property in user package
