+++ finexblock/entity/user/user_starred_trading_pair.model.go
@@ -1,9 +1,9 @@
 package user
 
 type UserStarredTraddingPair struct {
-	ID            uint `gorm:"primaryKey;autoIncrement:true;comment:'기본키'" json:"id"`
-	UserID        uint `json:"user_id" gorm:"comment:'유저 id'"`
-	OrderSymbolId uint `json:"order_symbol_id" gorm:"comment: '거래쌍 id'"`
+	ID            uint `gorm:"primaryKey;autoIncrement:true;comment:'기본키'"`
+	UserID        uint `gorm:"comment:'유저 id'"`
+	OrderSymbolId uint `gorm:"comment: '거래쌍 id'"`
 }
 
 func (s *UserStarredTraddingPair) Alias() string {
